// apps/core/src/modules/meta-preset/meta-preset.enum.ts
export enum MetaFieldType {
  Text = 'text',
  Textarea = 'textarea',
  Number = 'number',
  Url = 'url',
  Select = 'select',
  MultiSelect = 'multi-select',
  Checkbox = 'checkbox',
  Tags = 'tags',
  Boolean = 'boolean',
  Object = 'object',
}

export enum MetaPresetScope {
  Post = 'post',
  Note = 'note',
  Both = 'both',
}